python_fastapi/requirements.txt
fastapi==0.115.7
pydantic==2.10.6
python-dotenv==1.0.1
httpx==0.27.2
uvicorn==0.33.0
PyYAML==6.0.2
python-docx==1.1.2
python-jose==3.4.0
starlette==0.41.3
aisuite[all]
docstring_parser==0.16